Reorganize Benchmarks to follow project structure
