+++ benchmarks/Ardalis.Extensions.Benchmarks/Program.cs
@@ -1,13 +1,12 @@
-using System;
 using BenchmarkDotNet.Running;
+using Ardalis.Extensions.Benchmarks.StringManipulation;
 
-namespace Ardalis.Extensions.Benchmarks
+namespace Ardalis.Extensions.Benchmarks;
+
+public class Program
 {
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var summary = BenchmarkRunner.Run<Benchmarks>();
-        }
-    }
+  public static void Main(string[] args)
+  {
+    var summary = BenchmarkRunner.Run<LinesBenchmarks>();
+  }
 }
