+++ src/Ardalis.Extensions/StringManipulation/Lines.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Ardalis.Extensions.StringManipulation;
